refactor: modularize automated workflows
