--- src/core/refactoring/automated_refactoring_workflows.py
@@ -1,749 +1,0 @@
-#!/usr/bin/env python3
-"""
-Automated Refactoring Workflows - Agent-5
-=========================================
-
-This module implements automated workflows for refactoring operations as part of
-contract REFACTOR-002 deliverables.
-
-Features:
-- Automated workflow design and execution
-- Workflow automation with intelligent routing
-- Validation system for workflow reliability
-- Comprehensive testing and reliability metrics
-
-Author: Agent-5 (REFACTORING MANAGER)
-Contract: REFACTOR-002 - Automated Refactoring Workflow Implementation
-Status: IN PROGRESS
-"""
-
-import os
-import sys
-import json
-import logging
-import asyncio
-from pathlib import Path
-from typing import Dict, List, Any, Optional, Callable
-from dataclasses import dataclass, field
-from enum import Enum
-from datetime import datetime, timedelta
-import traceback
-
-# Add src to path for imports
-sys.path.append(str(Path(__file__).parent.parent.parent))
-
-from core.managers.base_manager import BaseManager
-from core.refactoring.advanced_refactoring_toolkit import AdvancedRefactoringToolkit
-from core.refactoring.refactoring_performance_benchmark import RefactoringPerformanceBenchmark
-from core.refactoring.automated_workflow_orchestrator import AutomatedWorkflowOrchestrator
-
-# Configure logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-
-class WorkflowStatus(Enum):
-    """Workflow execution status enumeration."""
-    PENDING = "pending"
-    RUNNING = "running"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    VALIDATED = "validated"
-
-
-class WorkflowType(Enum):
-    """Types of automated refactoring workflows."""
-    CODE_DUPLICATION_REMOVAL = "code_duplication_removal"
-    MONOLITH_BREAKDOWN = "monolith_breakdown"
-    SRP_VIOLATION_FIX = "srp_violation_fix"
-    PERFORMANCE_OPTIMIZATION = "performance_optimization"
-    ARCHITECTURE_CONSOLIDATION = "architecture_consolidation"
-    TEST_COVERAGE_IMPROVEMENT = "test_coverage_improvement"
-
-
-@dataclass
-class WorkflowStep:
-    """Individual step within a refactoring workflow."""
-    step_id: str
-    name: str
-    description: str
-    action: Callable
-    dependencies: List[str] = field(default_factory=list)
-    estimated_time: float = 0.0  # in minutes
-    status: WorkflowStatus = WorkflowStatus.PENDING
-    result: Any = None
-    error: Optional[str] = None
-    start_time: Optional[datetime] = None
-    end_time: Optional[datetime] = None
-
-
-@dataclass
-class WorkflowExecution:
-    """Complete workflow execution record."""
-    workflow_id: str
-    workflow_type: WorkflowType
-    target_files: List[str]
-    steps: List[WorkflowStep]
-    status: WorkflowStatus = WorkflowStatus.PENDING
-    start_time: Optional[datetime] = None
-    end_time: Optional[datetime] = None
-    total_time: Optional[float] = None
-    success_rate: float = 0.0
-    validation_results: Dict[str, Any] = field(default_factory=dict)
-
-
-class AutomatedRefactoringWorkflows(BaseManager):
-    """
-    Automated refactoring workflows implementation.
-    
-    This class provides comprehensive workflow automation for refactoring operations,
-    including intelligent routing, validation, and reliability testing.
-    """
-    
-    def __init__(self, config: Optional[Dict[str, Any]] = None):
-        """Initialize the automated refactoring workflows system."""
-        super().__init__(config or {})
-        self.workflows: Dict[str, WorkflowExecution] = {}
-        self.workflow_templates: Dict[WorkflowType, List[Dict[str, Any]]] = {}
-        self.validation_rules: Dict[str, Callable] = {}
-        self.reliability_metrics: Dict[str, float] = {}
-        
-        self._initialize_workflow_templates()
-        self._initialize_validation_rules()
-        self._setup_logging()
-    
-    def _initialize_workflow_templates(self):
-        """Initialize predefined workflow templates for different refactoring types."""
-        self.workflow_templates = {
-            WorkflowType.CODE_DUPLICATION_REMOVAL: [
-                {
-                    "step_id": "analysis",
-                    "name": "Code Duplication Analysis",
-                    "description": "Analyze codebase for duplicate code patterns",
-                    "estimated_time": 15.0
-                },
-                {
-                    "step_id": "identification",
-                    "name": "Duplicate Pattern Identification",
-                    "description": "Identify specific duplicate code sections",
-                    "estimated_time": 10.0
-                },
-                {
-                    "step_id": "extraction",
-                    "name": "Common Code Extraction",
-                    "description": "Extract common code into reusable functions",
-                    "estimated_time": 20.0
-                },
-                {
-                    "step_id": "refactoring",
-                    "name": "Duplicate Code Refactoring",
-                    "description": "Refactor code to use extracted functions",
-                    "estimated_time": 25.0
-                },
-                {
-                    "step_id": "validation",
-                    "name": "Refactoring Validation",
-                    "description": "Validate that refactoring maintains functionality",
-                    "estimated_time": 15.0
-                }
-            ],
-            WorkflowType.MONOLITH_BREAKDOWN: [
-                {
-                    "step_id": "analysis",
-                    "name": "Monolith Analysis",
-                    "description": "Analyze monolithic file structure and dependencies",
-                    "estimated_time": 20.0
-                },
-                {
-                    "step_id": "planning",
-                    "name": "Breakdown Planning",
-                    "description": "Plan modular breakdown strategy",
-                    "estimated_time": 15.0
-                },
-                {
-                    "step_id": "extraction",
-                    "name": "Module Extraction",
-                    "description": "Extract logical modules from monolith",
-                    "estimated_time": 30.0
-                },
-                {
-                    "step_id": "interface",
-                    "name": "Interface Definition",
-                    "description": "Define clean interfaces between modules",
-                    "estimated_time": 20.0
-                },
-                {
-                    "step_id": "testing",
-                    "name": "Module Testing",
-                    "description": "Test extracted modules for functionality",
-                    "estimated_time": 25.0
-                }
-            ],
-            WorkflowType.SRP_VIOLATION_FIX: [
-                {
-                    "step_id": "violation_analysis",
-                    "name": "SRP Violation Analysis",
-                    "description": "Identify Single Responsibility Principle violations",
-                    "estimated_time": 15.0
-                },
-                {
-                    "step_id": "responsibility_separation",
-                    "name": "Responsibility Separation",
-                    "description": "Separate mixed responsibilities into focused classes",
-                    "estimated_time": 25.0
-                },
-                {
-                    "step_id": "class_restructuring",
-                    "name": "Class Restructuring",
-                    "description": "Restructure classes to follow SRP",
-                    "estimated_time": 20.0
-                },
-                {
-                    "step_id": "dependency_management",
-                    "name": "Dependency Management",
-                    "description": "Manage dependencies between separated classes",
-                    "estimated_time": 15.0
-                },
-                {
-                    "step_id": "validation",
-                    "name": "SRP Compliance Validation",
-                    "description": "Validate that classes now follow SRP",
-                    "estimated_time": 15.0
-                }
-            ]
-        }
-    
-    def _initialize_validation_rules(self):
-        """Initialize validation rules for workflow reliability."""
-        self.validation_rules = {
-            "file_size_reduction": self._validate_file_size_reduction,
-            "function_count_optimization": self._validate_function_count_optimization,
-            "complexity_reduction": self._validate_complexity_reduction,
-            "test_coverage_maintenance": self._validate_test_coverage_maintenance,
-            "functionality_preservation": self._validate_functionality_preservation
-        }
-    
-    def create_workflow(self, workflow_type: WorkflowType, target_files: List[str], 
-                       custom_steps: Optional[List[Dict[str, Any]]] = None) -> str:
-        """
-        Create a new automated refactoring workflow.
-        
-        Args:
-            workflow_type: Type of refactoring workflow to create
-            target_files: List of target files for refactoring
-            custom_steps: Optional custom workflow steps
-            
-        Returns:
-            Workflow ID for tracking
-        """
-        workflow_id = f"workflow_{workflow_type.value}_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
-        
-        # Get template steps or use custom steps
-        template_steps = self.workflow_templates.get(workflow_type, [])
-        steps_data = custom_steps if custom_steps else template_steps
-        
-        # Create workflow steps
-        steps = []
-        for step_data in steps_data:
-            step = WorkflowStep(
-                step_id=step_data["step_id"],
-                name=step_data["name"],
-                description=step_data["description"],
-                action=self._get_step_action(step_data["step_id"]),
-                estimated_time=step_data["estimated_time"]
-            )
-            steps.append(step)
-        
-        # Create workflow execution
-        workflow = WorkflowExecution(
-            workflow_id=workflow_id,
-            workflow_type=workflow_type,
-            target_files=target_files,
-            steps=steps
-        )
-        
-        self.workflows[workflow_id] = workflow
-        logger.info(f"Created workflow {workflow_id} for {workflow_type.value}")
-        
-        return workflow_id
-    
-    def _get_step_action(self, step_id: str) -> Callable:
-        """Get the appropriate action function for a workflow step."""
-        actions = {
-            "analysis": self._execute_analysis_step,
-            "identification": self._execute_identification_step,
-            "extraction": self._execute_extraction_step,
-            "refactoring": self._execute_refactoring_step,
-            "validation": self._execute_validation_step,
-            "planning": self._execute_planning_step,
-            "interface": self._execute_interface_step,
-            "testing": self._execute_testing_step,
-            "violation_analysis": self._execute_violation_analysis_step,
-            "responsibility_separation": self._execute_responsibility_separation_step,
-            "class_restructuring": self._execute_class_restructuring_step,
-            "dependency_management": self._execute_dependency_management_step
-        }
-        return actions.get(step_id, self._execute_generic_step)
-    
-    async def execute_workflow(self, workflow_id: str) -> WorkflowExecution:
-        """
-        Execute a complete refactoring workflow.
-        
-        Args:
-            workflow_id: ID of the workflow to execute
-            
-        Returns:
-            Completed workflow execution record
-        """
-        if workflow_id not in self.workflows:
-            raise ValueError(f"Workflow {workflow_id} not found")
-        
-        workflow = self.workflows[workflow_id]
-        workflow.status = WorkflowStatus.RUNNING
-        workflow.start_time = datetime.now()
-        
-        logger.info(f"Starting workflow execution: {workflow_id}")
-        
-        try:
-            # Execute steps in dependency order
-            for step in workflow.steps:
-                await self._execute_workflow_step(workflow, step)
-            
-            workflow.status = WorkflowStatus.COMPLETED
-            workflow.end_time = datetime.now()
-            workflow.total_time = (workflow.end_time - workflow.start_time).total_seconds() / 60.0
-            
-            # Calculate success rate
-            completed_steps = [s for s in workflow.steps if s.status == WorkflowStatus.COMPLETED]
-            workflow.success_rate = len(completed_steps) / len(workflow.steps) * 100
-            
-            logger.info(f"Workflow {workflow_id} completed successfully in {workflow.total_time:.2f} minutes")
-            
-        except Exception as e:
-            workflow.status = WorkflowStatus.FAILED
-            workflow.end_time = datetime.now()
-            logger.error(f"Workflow {workflow_id} failed: {str(e)}")
-            logger.error(traceback.format_exc())
-        
-        return workflow
-    
-    async def _execute_workflow_step(self, workflow: WorkflowExecution, step: WorkflowStep):
-        """Execute a single workflow step."""
-        step.status = WorkflowStatus.RUNNING
-        step.start_time = datetime.now()
-        
-        logger.info(f"Executing step: {step.name}")
-        
-        try:
-            # Execute the step action
-            step.result = await step.action(workflow, step)
-            step.status = WorkflowStatus.COMPLETED
-            step.end_time = datetime.now()
-            
-            logger.info(f"Step {step.name} completed successfully")
-            
-        except Exception as e:
-            step.status = WorkflowStatus.FAILED
-            step.error = str(e)
-            step.end_time = datetime.now()
-            logger.error(f"Step {step.name} failed: {str(e)}")
-            raise
-    
-    async def _execute_analysis_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute analysis step for code analysis."""
-        # Simulate analysis execution
-        await asyncio.sleep(2)  # Simulate processing time
-        
-        analysis_results = {
-            "target_files": workflow.target_files,
-            "analysis_type": workflow.workflow_type.value,
-            "findings": f"Analysis completed for {len(workflow.target_files)} files",
-            "recommendations": "Proceed with next step"
-        }
-        
-        return analysis_results
-    
-    async def _execute_identification_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute identification step for pattern identification."""
-        await asyncio.sleep(1.5)
-        
-        identification_results = {
-            "patterns_found": 5,
-            "duplicate_sections": 3,
-            "refactoring_opportunities": 2
-        }
-        
-        return identification_results
-    
-    async def _execute_extraction_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute extraction step for code extraction."""
-        await asyncio.sleep(2.5)
-        
-        extraction_results = {
-            "extracted_functions": 3,
-            "new_modules_created": 1,
-            "code_reduction": "15%"
-        }
-        
-        return extraction_results
-    
-    async def _execute_refactoring_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute refactoring step for code refactoring."""
-        await asyncio.sleep(3.0)
-        
-        refactoring_results = {
-            "files_refactored": len(workflow.target_files),
-            "lines_modified": 45,
-            "refactoring_score": "85%"
-        }
-        
-        return refactoring_results
-    
-    async def _execute_validation_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute validation step for refactoring validation."""
-        await asyncio.sleep(1.5)
-        
-        validation_results = {
-            "tests_passed": 12,
-            "functionality_verified": True,
-            "quality_score": "92%"
-        }
-        
-        return validation_results
-    
-    async def _execute_planning_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute planning step for breakdown planning."""
-        await asyncio.sleep(2.0)
-        
-        planning_results = {
-            "modules_planned": 4,
-            "dependencies_mapped": True,
-            "breakdown_strategy": "Logical separation by responsibility"
-        }
-        
-        return planning_results
-    
-    async def _execute_interface_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute interface step for interface definition."""
-        await asyncio.sleep(2.5)
-        
-        interface_results = {
-            "interfaces_defined": 3,
-            "api_contracts": 2,
-            "coupling_reduced": "40%"
-        }
-        
-        return interface_results
-    
-    async def _execute_testing_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute testing step for module testing."""
-        await asyncio.sleep(2.0)
-        
-        testing_results = {
-            "test_cases": 18,
-            "coverage": "87%",
-            "all_tests_passed": True
-        }
-        
-        return testing_results
-    
-    async def _execute_violation_analysis_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute SRP violation analysis step."""
-        await asyncio.sleep(2.0)
-        
-        violation_results = {
-            "srp_violations_found": 4,
-            "mixed_responsibilities": 3,
-            "refactoring_priority": "HIGH"
-        }
-        
-        return violation_results
-    
-    async def _execute_responsibility_separation_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute responsibility separation step."""
-        await asyncio.sleep(3.0)
-        
-        separation_results = {
-            "responsibilities_separated": 4,
-            "new_classes_created": 3,
-            "separation_score": "88%"
-        }
-        
-        return separation_results
-    
-    async def _execute_class_restructuring_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute class restructuring step."""
-        await asyncio.sleep(2.5)
-        
-        restructuring_results = {
-            "classes_restructured": 3,
-            "methods_reorganized": 8,
-            "structure_improvement": "85%"
-        }
-        
-        return restructuring_results
-    
-    async def _execute_dependency_management_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute dependency management step."""
-        await asyncio.sleep(2.0)
-        
-        dependency_results = {
-            "dependencies_managed": 5,
-            "coupling_reduced": "35%",
-            "cohesion_improved": "90%"
-        }
-        
-        return dependency_results
-    
-    async def _execute_generic_step(self, workflow: WorkflowExecution, step: WorkflowStep) -> Dict[str, Any]:
-        """Execute generic step for undefined step types."""
-        await asyncio.sleep(1.0)
-        
-        generic_results = {
-            "step_type": "generic",
-            "status": "completed",
-            "message": "Generic step execution completed"
-        }
-        
-        return generic_results
-    
-    def validate_workflow(self, workflow_id: str) -> Dict[str, Any]:
-        """
-        Validate a completed workflow for reliability and quality.
-        
-        Args:
-            workflow_id: ID of the workflow to validate
-            
-        Returns:
-            Validation results and reliability metrics
-        """
-        if workflow_id not in self.workflows:
-            raise ValueError(f"Workflow {workflow_id} not found")
-        
-        workflow = self.workflows[workflow_id]
-        
-        if workflow.status != WorkflowStatus.COMPLETED:
-            raise ValueError(f"Workflow {workflow_id} is not completed")
-        
-        validation_results = {}
-        
-        # Run all validation rules
-        for rule_name, rule_func in self.validation_rules.items():
-            try:
-                validation_results[rule_name] = rule_func(workflow)
-            except Exception as e:
-                validation_results[rule_name] = {"status": "error", "error": str(e)}
-        
-        # Calculate overall reliability score
-        successful_validations = sum(1 for result in validation_results.values() 
-                                   if isinstance(result, dict) and result.get("status") == "passed")
-        total_validations = len(validation_results)
-        reliability_score = (successful_validations / total_validations) * 100 if total_validations > 0 else 0
-        
-        validation_results["overall_reliability"] = {
-            "score": reliability_score,
-            "status": "passed" if reliability_score >= 80 else "failed",
-            "total_validations": total_validations,
-            "successful_validations": successful_validations
-        }
-        
-        # Update workflow with validation results
-        workflow.validation_results = validation_results
-        workflow.status = WorkflowStatus.VALIDATED
-        
-        # Store reliability metrics
-        self.reliability_metrics[workflow_id] = reliability_score
-        
-        logger.info(f"Workflow {workflow_id} validated with reliability score: {reliability_score:.2f}%")
-        
-        return validation_results
-    
-    def _validate_file_size_reduction(self, workflow: WorkflowExecution) -> Dict[str, Any]:
-        """Validate that file sizes were reduced during refactoring."""
-        # Simulate file size validation
-        return {
-            "status": "passed",
-            "message": "File size reduction validation passed",
-            "details": "Target files reduced by 25-40%"
-        }
-    
-    def _validate_function_count_optimization(self, workflow: WorkflowExecution) -> Dict[str, Any]:
-        """Validate that function counts were optimized."""
-        return {
-            "status": "passed",
-            "message": "Function count optimization validation passed",
-            "details": "Functions properly organized and optimized"
-        }
-    
-    def _validate_complexity_reduction(self, workflow: WorkflowExecution) -> Dict[str, Any]:
-        """Validate that code complexity was reduced."""
-        return {
-            "status": "passed",
-            "message": "Complexity reduction validation passed",
-            "details": "Cyclomatic complexity reduced by 30%"
-        }
-    
-    def _validate_test_coverage_maintenance(self, workflow: WorkflowExecution) -> Dict[str, Any]:
-        """Validate that test coverage was maintained or improved."""
-        return {
-            "status": "passed",
-            "message": "Test coverage maintenance validation passed",
-            "details": "Test coverage maintained at 85%+"
-        }
-    
-    def _validate_functionality_preservation(self, workflow: WorkflowExecution) -> Dict[str, Any]:
-        """Validate that all functionality was preserved during refactoring."""
-        return {
-            "status": "passed",
-            "message": "Functionality preservation validation passed",
-            "details": "All core functionality maintained"
-        }
-    
-    def get_workflow_status(self, workflow_id: str) -> Optional[WorkflowExecution]:
-        """Get the current status of a workflow."""
-        return self.workflows.get(workflow_id)
-    
-    def list_workflows(self) -> List[Dict[str, Any]]:
-        """List all workflows with their current status."""
-        workflow_list = []
-        
-        for workflow_id, workflow in self.workflows.items():
-            workflow_info = {
-                "workflow_id": workflow_id,
-                "type": workflow.workflow_type.value,
-                "status": workflow.status.value,
-                "target_files": len(workflow.target_files),
-                "steps": len(workflow.steps),
-                "success_rate": workflow.success_rate,
-                "start_time": workflow.start_time.isoformat() if workflow.start_time else None,
-                "total_time": workflow.total_time
-            }
-            workflow_list.append(workflow_info)
-        
-        return workflow_list
-    
-    def get_reliability_metrics(self) -> Dict[str, float]:
-        """Get reliability metrics for all workflows."""
-        return self.reliability_metrics.copy()
-    
-    def cleanup_completed_workflows(self, max_age_days: int = 7):
-        """Clean up old completed workflows to free memory."""
-        cutoff_date = datetime.now() - timedelta(days=max_age_days)
-        workflows_to_remove = []
-        
-        for workflow_id, workflow in self.workflows.items():
-            if (workflow.status in [WorkflowStatus.COMPLETED, WorkflowStatus.VALIDATED] and
-                workflow.end_time and workflow.end_time < cutoff_date):
-                workflows_to_remove.append(workflow_id)
-        
-        for workflow_id in workflows_to_remove:
-            del self.workflows[workflow_id]
-            logger.info(f"Cleaned up old workflow: {workflow_id}")
-    
-    def export_workflow_report(self, workflow_id: str, output_path: str) -> bool:
-        """
-        Export a workflow execution report to JSON.
-        
-        Args:
-            workflow_id: ID of the workflow to export
-            output_path: Path to save the report
-            
-        Returns:
-            True if export successful, False otherwise
-        """
-        if workflow_id not in self.workflows:
-            return False
-        
-        workflow = self.workflows[workflow_id]
-        
-        # Prepare export data
-        export_data = {
-            "workflow_id": workflow.workflow_id,
-            "workflow_type": workflow.workflow_type.value,
-            "target_files": workflow.target_files,
-            "status": workflow.status.value,
-            "start_time": workflow.start_time.isoformat() if workflow.start_time else None,
-            "end_time": workflow.end_time.isoformat() if workflow.end_time else None,
-            "total_time": workflow.total_time,
-            "success_rate": workflow.success_rate,
-            "steps": [
-                {
-                    "step_id": step.step_id,
-                    "name": step.name,
-                    "description": step.description,
-                    "status": step.status.value,
-                    "result": step.result,
-                    "error": step.error,
-                    "start_time": step.start_time.isoformat() if step.start_time else None,
-                    "end_time": step.end_time.isoformat() if step.end_time else None
-                }
-                for step in workflow.steps
-            ],
-            "validation_results": workflow.validation_results
-        }
-        
-        try:
-            with open(output_path, 'w') as f:
-                json.dump(export_data, f, indent=2, default=str)
-            
-            logger.info(f"Workflow report exported to: {output_path}")
-            return True
-            
-        except Exception as e:
-            logger.error(f"Failed to export workflow report: {str(e)}")
-            return False
-
-
-# Example usage and testing
-async def demo_automated_workflows():
-    """Demonstrate the automated refactoring workflows system."""
-    print("🚀 Automated Refactoring Workflows Demo")
-    print("=" * 50)
-    
-    # Initialize workflows system
-    workflows = AutomatedRefactoringWorkflows()
-    
-    # Create a workflow for code duplication removal
-    target_files = [
-        "src/services/financial/portfolio/rebalancing.py",
-        "src/core/performance/performance_orchestrator.py"
-    ]
-    
-    workflow_id = workflows.create_workflow(
-        WorkflowType.CODE_DUPLICATION_REMOVAL,
-        target_files
-    )
-    
-    print(f"✅ Created workflow: {workflow_id}")
-    
-    # Execute the workflow
-    print("\n🔄 Executing workflow...")
-    workflow = await workflows.execute_workflow(workflow_id)
-    
-    print(f"✅ Workflow completed with {workflow.success_rate:.1f}% success rate")
-    print(f"⏱️  Total execution time: {workflow.total_time:.2f} minutes")
-    
-    # Validate the workflow
-    print("\n🔍 Validating workflow...")
-    validation_results = workflows.validate_workflow(workflow_id)
-    
-    reliability_score = validation_results.get("overall_reliability", {}).get("score", 0)
-    print(f"✅ Workflow validated with {reliability_score:.1f}% reliability")
-    
-    # Export workflow report
-    report_path = "workflow_execution_report.json"
-    if workflows.export_workflow_report(workflow_id, report_path):
-        print(f"📊 Workflow report exported to: {report_path}")
-    
-    # List all workflows
-    print("\n📋 All workflows:")
-    for workflow_info in workflows.list_workflows():
-        print(f"  - {workflow_info['workflow_id']}: {workflow_info['status']} "
-              f"({workflow_info['success_rate']:.1f}% success)")
-    
-    print("\n🎉 Demo completed successfully!")
-
-
-if __name__ == "__main__":
-    # Run the demo
-    asyncio.run(demo_automated_workflows())
